Major changes to teh Jaccard index resampling function.
Fix an error in sampling and calculation the Jaccard index from the samples.
Sample only from the null set and use two matrices to hold the intersect
and the union. Rename some function arguments to make it more consistent and
add comments. Add benchmark to compare vectorized version of resampling function
with normal sapply.